Added 1d cnn model to prediction options.  Rearranged functions from view to helper.
